// apps/web/src/components/motion/FadeIn.d.ts
import React from 'react';
interface FadeInProps {
    children: React.ReactNode;
    className?: string;
    delay?: number;
    duration?: number;
}
export declare const FadeIn: ({ children, duration, delay, className, }: FadeInProps) => import("react/jsx-runtime").JSX.Element;
export {};
//# sourceMappingURL=FadeIn.d.ts.map